% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setlists.R
\docType{data}
\name{setlists}
\alias{setlists}
\title{Metadata for setlists in the spRingsteen dataset.}
\format{
A data frame with 52,100 rows and 4 variables:
\describe{
 \item{gig_key}{key associated with the concert which the setlist is from.}
 \item{song_key}{key associated with the song played.}
 \item{song}{name of the song played.}
 \item{song_number}{order of appearance for the song in the setlist.}
 }
}
\usage{
setlists
}
\description{
Metadata for the setlists of concerts played by Bruce Springsteen
both solo and with numerous bands from the years 1973 to 2021.
}
\examples{

library(dplyr)
# what are the top five most played songs?

setlists \%>\%
  count(song, sort = TRUE) \%>\%
  slice(1:5)

# what is the average show length?

setlists \%>\%
  count(gig_key) \%>\%
  summarise(ave_length = mean(n))

}
\seealso{
\url{http://brucebase.wikidot.com/}
}
\keyword{datasets}
